mod: 요약, 차트 컴포넌트화
